anno/requirements.txt
absl-py==1.3.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asone==0.1.2.dev17
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==6.0.0
cachetools==5.2.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi==1.15.1
charset-normalizer==2.1.1
coloredlogs==15.0.1
comm==0.1.2
commonmark==0.9.1
contourpy==1.0.6
cycler==0.11.0
Cython==0.29.33
cython-bbox==0.1.3
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
easydict==1.10
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.9.0
filterpy==1.4.5
flatbuffers==23.1.21
fonttools==4.38.0
fqdn==1.5.1
gdown==4.6.0
google-auth==2.15.0
google-auth-oauthlib==0.4.6
grpcio==1.51.1
humanfriendly==10.0
idna==3.4
imageio==2.22.4
importlib-metadata==5.1.0
importlib-resources==5.10.2
ipykernel==6.20.2
ipython==8.7.0
ipython-genutils==0.2.0
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter_client==8.0.1
jupyter_core==5.1.5
jupyter_server==2.1.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
keyboard==0.13.5
kiwisolver==1.4.4
lap==0.4.0
loguru==0.6.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mistune==2.0.4
mpmath==1.2.1
nbclassic==0.5.1
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==2.8.8
norfair==2.2.0
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
onnxruntime-gpu==1.12.1
opencv-python==4.7.0.68
packaging==21.3
pandas==1.5.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pkgutil_resolve_name==1.3.10
platformdirs==2.6.2
prometheus-client==0.16.0
prompt-toolkit==3.0.33
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.4
pytz==2022.6
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.0
requests==2.28.1
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==12.0.1
rsa==4.9
scikit-image==0.19.3
scipy==1.9.3
seaborn==0.12.1
Send2Trash==1.8.0
sentry-sdk==1.14.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.6.2
sympy==1.11.1
tabulate==0.9.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.17.1
thop==0.1.1.post2209072238
tifffile==2022.10.10
tinycss2==1.2.1
torch==1.13.1
torchvision==0.14.1
tornado==6.2
tqdm==4.64.1
traitlets==5.6.0
typing-extensions==3.10.0.2
ultralytics==8.0.21
uri-template==1.2.0
urllib3==1.26.14
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.0
Werkzeug==2.2.2
zipp==3.11.0
